+++ brewtils/test/comparable.py
@@ -14,6 +14,7 @@ from functools import partial
 
 import brewtils.test
 from brewtils.models import (
+    AliasUserMap,
     Choices,
     Command,
     Connection,
@@ -24,21 +25,26 @@ from brewtils.models import (
     Instance,
     IntervalTrigger,
     Job,
-    LegacyRole,
     LoggingConfig,
     Operation,
     Parameter,
     PatchOperation,
-    Principal,
     Queue,
+    Replication,
     Request,
     RequestFile,
     RequestTemplate,
     Resolvable,
+    Role,
     Runner,
-    System,
+    StatusHistory,
+    StatusInfo,
     Subscriber,
+    System,
     Topic,
+    UpstreamRole,
+    User,
+    UserToken,
 )
 
 __all__ = [
@@ -52,7 +58,9 @@ __all__ = [
     "assert_trigger_equal",
     "assert_command_equal",
     "assert_parameter_equal",
-    "assert_principal_equal",
+    "assert_user_token_equal",
+    "assert_user_equal",
+    "assert_alias_user_map_equal",
     "assert_request_equal",
     "assert_role_equal",
     "assert_system_equal",
@@ -62,6 +70,9 @@ __all__ = [
     "assert_runner_equal",
     "assert_subscriber_equal",
     "assert_topic_equal",
+    "assert_status_info_equal",
+    "assert_status_history_equal",
+    "assert_replication_equal",
 ]
 
 
@@ -184,7 +195,6 @@ def _assert_wrapper(obj1, obj2, expected_type=None, do_raise=False, **kwargs):
 
 
 # These are the 'simple' models - they don't have any nested models as fields
-assert_instance_equal = partial(_assert_wrapper, expected_type=Instance)
 assert_choices_equal = partial(_assert_wrapper, expected_type=Choices)
 assert_patch_equal = partial(_assert_wrapper, expected_type=PatchOperation)
 assert_logging_config_equal = partial(_assert_wrapper, expected_type=LoggingConfig)
@@ -196,8 +206,45 @@ assert_trigger_equal = partial(
 assert_request_file_equal = partial(_assert_wrapper, expected_type=RequestFile)
 assert_runner_equal = partial(_assert_wrapper, expected_type=Runner)
 assert_resolvable_equal = partial(_assert_wrapper, expected_type=Resolvable)
-assert_connection_equal = partial(_assert_wrapper, expected_type=Connection)
 assert_subscriber_equal = partial(_assert_wrapper, expected_type=Subscriber)
+assert_status_history_equal = partial(_assert_wrapper, expected_type=StatusHistory)
+
+
+def assert_status_info_equal(obj1, obj2, do_raise=False):
+    return _assert_wrapper(
+        obj1,
+        obj2,
+        expected_type=StatusInfo,
+        deep_fields={
+            "history": partial(assert_status_history_equal, do_raise=True),
+        },
+        do_raise=do_raise,
+    )
+
+
+def assert_instance_equal(obj1, obj2, do_raise=False):
+    return _assert_wrapper(
+        obj1,
+        obj2,
+        expected_type=Instance,
+        deep_fields={"status_info": partial(assert_status_info_equal, do_raise=True)},
+        do_raise=do_raise,
+    )
+
+
+def assert_connection_equal(obj1, obj2, do_raise=False):
+    return _assert_wrapper(
+        obj1,
+        obj2,
+        expected_type=Connection,
+        deep_fields={"status_info": partial(assert_status_info_equal, do_raise=True)},
+        do_raise=do_raise,
+    )
+
+
+assert_alias_user_map_equal = partial(_assert_wrapper, expected_type=AliasUserMap)
+assert_subscriber_equal = partial(_assert_wrapper, expected_type=Subscriber)
+assert_replication_equal = partial(_assert_wrapper, expected_type=Replication)
 
 
 def assert_command_equal(obj1, obj2, do_raise=False):
@@ -241,12 +288,28 @@ def assert_event_equal(obj1, obj2, do_raise=False):
     )
 
 
-def assert_principal_equal(obj1, obj2, do_raise=False):
+def assert_user_equal(obj1, obj2, do_raise=False):
     return _assert_wrapper(
         obj1,
         obj2,
-        expected_type=Principal,
-        deep_fields={"roles": partial(assert_role_equal, do_raise=True)},
+        expected_type=User,
+        deep_fields={
+            "local_roles": partial(assert_role_equal, do_raise=True),
+            "upstream_roles": partial(assert_upstream_role_equal, do_raise=True),
+            "user_alias_mapping": partial(assert_alias_user_map_equal, do_raise=True),
+        },
+        do_raise=do_raise,
+    )
+
+
+def assert_user_token_equal(obj1, obj2, do_raise=False):
+    return _assert_wrapper(
+        obj1,
+        obj2,
+        expected_type=UserToken,
+        deep_fields={
+            "user": partial(assert_user_equal, do_raise=True),
+        },
         do_raise=do_raise,
     )
 
@@ -303,8 +366,16 @@ def assert_role_equal(obj1, obj2, do_raise=False):
     return _assert_wrapper(
         obj1,
         obj2,
-        expected_type=LegacyRole,
-        deep_fields={"roles": partial(assert_role_equal, do_raise=True)},
+        expected_type=Role,
+        do_raise=do_raise,
+    )
+
+
+def assert_upstream_role_equal(obj1, obj2, do_raise=False):
+    return _assert_wrapper(
+        obj1,
+        obj2,
+        expected_type=UpstreamRole,
         do_raise=do_raise,
     )
 
@@ -330,6 +401,7 @@ def assert_job_equal(obj1, obj2, do_raise=False):
         deep_fields={
             "trigger": partial(assert_trigger_equal, do_raise=True),
             "request_template": partial(assert_request_template_equal, do_raise=True),
+            "replication": partial(assert_replication_equal, do_raise=True),
         },
         do_raise=do_raise,
     )
@@ -390,6 +462,7 @@ def assert_garden_equal(obj1, obj2, do_raise=False):
             "systems": partial(assert_system_equal, do_raise=True),
             "receiving_connections": partial(assert_connection_equal, do_raise=True),
             "publishing_connections": partial(assert_connection_equal, do_raise=True),
+            "status_info": partial(assert_status_info_equal, do_raise=True),
         },
         do_raise=do_raise,
     )