+++ test/schema_parser_test.py
@@ -5,11 +5,15 @@ from __future__ import unicode_literals
 
 import copy
 
-import brewtils.models
 import pytest
+from marshmallow.exceptions import MarshmallowError
+from pytest_lazyfixture import lazy_fixture
+
+import brewtils.models
 from brewtils.models import System
 from brewtils.schema_parser import SchemaParser
 from brewtils.test.comparable import (
+    assert_alias_user_map_equal,
     assert_command_equal,
     assert_connection_equal,
     assert_event_equal,
@@ -21,19 +25,22 @@ from brewtils.test.comparable import (
     assert_operation_equal,
     assert_parameter_equal,
     assert_patch_equal,
-    assert_principal_equal,
     assert_queue_equal,
+    assert_replication_equal,
     assert_request_equal,
     assert_request_file_equal,
     assert_resolvable_equal,
     assert_role_equal,
     assert_runner_equal,
+    assert_status_history_equal,
+    assert_status_info_equal,
     assert_subscriber_equal,
     assert_system_equal,
     assert_topic_equal,
+    assert_upstream_role_equal,
+    assert_user_equal,
+    assert_user_token_equal,
 )
-from marshmallow.exceptions import MarshmallowError
-from pytest_lazyfixture import lazy_fixture
 
 
 class TestParse(object):
@@ -123,16 +130,34 @@ class TestParse(object):
                 lazy_fixture("bg_queue"),
             ),
             (
-                brewtils.models.Principal,
-                lazy_fixture("principal_dict"),
-                assert_principal_equal,
-                lazy_fixture("bg_principal"),
+                brewtils.models.User,
+                lazy_fixture("user_dict"),
+                assert_user_equal,
+                lazy_fixture("bg_user"),
             ),
             (
-                brewtils.models.LegacyRole,
-                lazy_fixture("legacy_role_dict"),
+                brewtils.models.UserToken,
+                lazy_fixture("user_token_dict"),
+                assert_user_token_equal,
+                lazy_fixture("bg_user_token"),
+            ),
+            (
+                brewtils.models.AliasUserMap,
+                lazy_fixture("alias_user_map_dict"),
+                assert_alias_user_map_equal,
+                lazy_fixture("bg_alias_user_map"),
+            ),
+            (
+                brewtils.models.Role,
+                lazy_fixture("role_dict"),
                 assert_role_equal,
                 lazy_fixture("bg_role"),
+            ),
+            (
+                brewtils.models.UpstreamRole,
+                lazy_fixture("upstream_role_dict"),
+                assert_upstream_role_equal,
+                lazy_fixture("bg_upstream_role"),
             ),
             (
                 brewtils.models.Job,
@@ -193,6 +218,24 @@ class TestParse(object):
                 lazy_fixture("topic_dict"),
                 assert_topic_equal,
                 lazy_fixture("bg_topic"),
+            ),
+            (
+                brewtils.models.StatusInfo,
+                lazy_fixture("status_info_dict"),
+                assert_status_info_equal,
+                lazy_fixture("bg_status_info"),
+            ),
+            (
+                brewtils.models.StatusHistory,
+                lazy_fixture("status_history_dict"),
+                assert_status_history_equal,
+                lazy_fixture("bg_status_history"),
+            ),
+            (
+                brewtils.models.Replication,
+                lazy_fixture("replication_dict"),
+                assert_replication_equal,
+                lazy_fixture("bg_replication"),
             ),
         ],
     )
@@ -263,16 +306,34 @@ class TestParse(object):
                 lazy_fixture("bg_queue"),
             ),
             (
-                "parse_principal",
-                lazy_fixture("principal_dict"),
-                assert_principal_equal,
-                lazy_fixture("bg_principal"),
+                "parse_user",
+                lazy_fixture("user_dict"),
+                assert_user_equal,
+                lazy_fixture("bg_user"),
+            ),
+            (
+                "parse_user_token",
+                lazy_fixture("user_token_dict"),
+                assert_user_token_equal,
+                lazy_fixture("bg_user_token"),
+            ),
+            (
+                "parse_alias_user_map",
+                lazy_fixture("alias_user_map_dict"),
+                assert_alias_user_map_equal,
+                lazy_fixture("bg_alias_user_map"),
             ),
             (
                 "parse_role",
-                lazy_fixture("legacy_role_dict"),
+                lazy_fixture("role_dict"),
                 assert_role_equal,
                 lazy_fixture("bg_role"),
+            ),
+            (
+                "parse_upstream_role",
+                lazy_fixture("upstream_role_dict"),
+                assert_upstream_role_equal,
+                lazy_fixture("bg_upstream_role"),
             ),
             (
                 "parse_job",
@@ -339,6 +400,24 @@ class TestParse(object):
                 lazy_fixture("topic_dict"),
                 assert_topic_equal,
                 lazy_fixture("bg_topic"),
+            ),
+            (
+                "parse_status_info",
+                lazy_fixture("status_info_dict"),
+                assert_status_info_equal,
+                lazy_fixture("bg_status_info"),
+            ),
+            (
+                "parse_status_history",
+                lazy_fixture("status_history_dict"),
+                assert_status_history_equal,
+                lazy_fixture("bg_status_history"),
+            ),
+            (
+                "parse_replication",
+                lazy_fixture("replication_dict"),
+                assert_replication_equal,
+                lazy_fixture("bg_replication"),
             ),
         ],
     )
@@ -408,16 +487,34 @@ class TestParse(object):
                 lazy_fixture("bg_queue"),
             ),
             (
-                brewtils.models.Principal,
-                lazy_fixture("principal_dict"),
-                assert_principal_equal,
-                lazy_fixture("bg_principal"),
+                brewtils.models.User,
+                lazy_fixture("user_dict"),
+                assert_user_equal,
+                lazy_fixture("bg_user"),
             ),
             (
-                brewtils.models.LegacyRole,
-                lazy_fixture("legacy_role_dict"),
+                brewtils.models.UserToken,
+                lazy_fixture("user_token_dict"),
+                assert_user_token_equal,
+                lazy_fixture("bg_user_token"),
+            ),
+            (
+                brewtils.models.AliasUserMap,
+                lazy_fixture("alias_user_map_dict"),
+                assert_alias_user_map_equal,
+                lazy_fixture("bg_alias_user_map"),
+            ),
+            (
+                brewtils.models.Role,
+                lazy_fixture("role_dict"),
                 assert_role_equal,
                 lazy_fixture("bg_role"),
+            ),
+            (
+                brewtils.models.UpstreamRole,
+                lazy_fixture("upstream_role_dict"),
+                assert_upstream_role_equal,
+                lazy_fixture("bg_upstream_role"),
             ),
             (
                 brewtils.models.Job,
@@ -478,6 +575,24 @@ class TestParse(object):
                 lazy_fixture("topic_dict"),
                 assert_topic_equal,
                 lazy_fixture("bg_topic"),
+            ),
+            (
+                brewtils.models.StatusInfo,
+                lazy_fixture("status_info_dict"),
+                assert_status_info_equal,
+                lazy_fixture("bg_status_info"),
+            ),
+            (
+                brewtils.models.StatusHistory,
+                lazy_fixture("status_history_dict"),
+                assert_status_history_equal,
+                lazy_fixture("bg_status_history"),
+            ),
+            (
+                brewtils.models.Replication,
+                lazy_fixture("replication_dict"),
+                assert_replication_equal,
+                lazy_fixture("bg_replication"),
             ),
         ],
     )
@@ -546,16 +661,34 @@ class TestParse(object):
                 lazy_fixture("bg_queue"),
             ),
             (
-                "parse_principal",
-                lazy_fixture("principal_dict"),
-                assert_principal_equal,
-                lazy_fixture("bg_principal"),
+                "parse_user",
+                lazy_fixture("user_dict"),
+                assert_user_equal,
+                lazy_fixture("bg_user"),
+            ),
+            (
+                "parse_user_token",
+                lazy_fixture("user_token_dict"),
+                assert_user_token_equal,
+                lazy_fixture("bg_user_token"),
+            ),
+            (
+                "parse_alias_user_map",
+                lazy_fixture("alias_user_map_dict"),
+                assert_alias_user_map_equal,
+                lazy_fixture("bg_alias_user_map"),
             ),
             (
                 "parse_role",
-                lazy_fixture("legacy_role_dict"),
+                lazy_fixture("role_dict"),
                 assert_role_equal,
                 lazy_fixture("bg_role"),
+            ),
+            (
+                "parse_upstream_role",
+                lazy_fixture("upstream_role_dict"),
+                assert_upstream_role_equal,
+                lazy_fixture("bg_upstream_role"),
             ),
             (
                 "parse_job",
@@ -611,6 +744,24 @@ class TestParse(object):
                 assert_topic_equal,
                 lazy_fixture("bg_topic"),
             ),
+            (
+                "parse_status_info",
+                lazy_fixture("status_info_dict"),
+                assert_status_info_equal,
+                lazy_fixture("bg_status_info"),
+            ),
+            (
+                "parse_status_history",
+                lazy_fixture("status_history_dict"),
+                assert_status_history_equal,
+                lazy_fixture("bg_status_history"),
+            ),
+            (
+                "parse_replication",
+                lazy_fixture("replication_dict"),
+                assert_replication_equal,
+                lazy_fixture("bg_replication"),
+            ),
         ],
     )
     def test_many_specific(self, method, data, assertion, expected):
@@ -655,8 +806,11 @@ class TestSerialize(object):
             (lazy_fixture("bg_logging_config"), lazy_fixture("logging_config_dict")),
             (lazy_fixture("bg_event"), lazy_fixture("event_dict")),
             (lazy_fixture("bg_queue"), lazy_fixture("queue_dict")),
-            (lazy_fixture("bg_principal"), lazy_fixture("principal_dict")),
-            (lazy_fixture("bg_role"), lazy_fixture("legacy_role_dict")),
+            (lazy_fixture("bg_user"), lazy_fixture("user_dict")),
+            (lazy_fixture("bg_user_token"), lazy_fixture("user_token_dict")),
+            (lazy_fixture("bg_alias_user_map"), lazy_fixture("alias_user_map_dict")),
+            (lazy_fixture("bg_role"), lazy_fixture("role_dict")),
+            (lazy_fixture("bg_upstream_role"), lazy_fixture("upstream_role_dict")),
             (lazy_fixture("bg_job"), lazy_fixture("job_dict")),
             (lazy_fixture("bg_cron_job"), lazy_fixture("cron_job_dict")),
             (lazy_fixture("bg_interval_job"), lazy_fixture("interval_job_dict")),
@@ -666,6 +820,9 @@ class TestSerialize(object):
             (lazy_fixture("bg_resolvable"), lazy_fixture("resolvable_dict")),
             (lazy_fixture("bg_subscriber"), lazy_fixture("subscriber_dict")),
             (lazy_fixture("bg_topic"), lazy_fixture("topic_dict")),
+            (lazy_fixture("bg_status_info"), lazy_fixture("status_info_dict")),
+            (lazy_fixture("bg_status_history"), lazy_fixture("status_history_dict")),
+            (lazy_fixture("bg_replication"), lazy_fixture("replication_dict")),
         ],
     )
     def test_single(self, model, expected):
@@ -717,14 +874,29 @@ class TestSerialize(object):
             ("serialize_event", lazy_fixture("bg_event"), lazy_fixture("event_dict")),
             ("serialize_queue", lazy_fixture("bg_queue"), lazy_fixture("queue_dict")),
             (
-                "serialize_principal",
-                lazy_fixture("bg_principal"),
-                lazy_fixture("principal_dict"),
+                "serialize_user",
+                lazy_fixture("bg_user"),
+                lazy_fixture("user_dict"),
+            ),
+            (
+                "serialize_user_token",
+                lazy_fixture("bg_user_token"),
+                lazy_fixture("user_token_dict"),
+            ),
+            (
+                "serialize_alias_user_map",
+                lazy_fixture("bg_alias_user_map"),
+                lazy_fixture("alias_user_map_dict"),
             ),
             (
                 "serialize_role",
                 lazy_fixture("bg_role"),
-                lazy_fixture("legacy_role_dict"),
+                lazy_fixture("role_dict"),
+            ),
+            (
+                "serialize_upstream_role",
+                lazy_fixture("bg_upstream_role"),
+                lazy_fixture("upstream_role_dict"),
             ),
             ("serialize_job", lazy_fixture("bg_job"), lazy_fixture("job_dict")),
             (
@@ -777,6 +949,21 @@ class TestSerialize(object):
                 lazy_fixture("bg_topic"),
                 lazy_fixture("topic_dict"),
             ),
+            (
+                "serialize_status_info",
+                lazy_fixture("bg_status_info"),
+                lazy_fixture("status_info_dict"),
+            ),
+            (
+                "serialize_status_history",
+                lazy_fixture("bg_status_history"),
+                lazy_fixture("status_history_dict"),
+            ),
+            (
+                "serialize_replication",
+                lazy_fixture("bg_replication"),
+                lazy_fixture("replication_dict"),
+            ),
         ],
     )
     def test_single_specific(self, method, data, expected):
@@ -796,8 +983,11 @@ class TestSerialize(object):
             (lazy_fixture("bg_logging_config"), lazy_fixture("logging_config_dict")),
             (lazy_fixture("bg_event"), lazy_fixture("event_dict")),
             (lazy_fixture("bg_queue"), lazy_fixture("queue_dict")),
-            (lazy_fixture("bg_principal"), lazy_fixture("principal_dict")),
-            (lazy_fixture("bg_role"), lazy_fixture("legacy_role_dict")),
+            (lazy_fixture("bg_user"), lazy_fixture("user_dict")),
+            (lazy_fixture("bg_user_token"), lazy_fixture("user_token_dict")),
+            (lazy_fixture("bg_alias_user_map"), lazy_fixture("alias_user_map_dict")),
+            (lazy_fixture("bg_role"), lazy_fixture("role_dict")),
+            (lazy_fixture("bg_upstream_role"), lazy_fixture("upstream_role_dict")),
             (lazy_fixture("bg_job"), lazy_fixture("job_dict")),
             (lazy_fixture("bg_cron_job"), lazy_fixture("cron_job_dict")),
             (lazy_fixture("bg_interval_job"), lazy_fixture("interval_job_dict")),
@@ -807,6 +997,9 @@ class TestSerialize(object):
             (lazy_fixture("bg_resolvable"), lazy_fixture("resolvable_dict")),
             (lazy_fixture("bg_subscriber"), lazy_fixture("subscriber_dict")),
             (lazy_fixture("bg_topic"), lazy_fixture("topic_dict")),
+            (lazy_fixture("bg_status_info"), lazy_fixture("status_info_dict")),
+            (lazy_fixture("bg_status_history"), lazy_fixture("status_history_dict")),
+            (lazy_fixture("bg_replication"), lazy_fixture("replication_dict")),
         ],
     )
     def test_many(self, model, expected):
@@ -875,11 +1068,26 @@ class TestRoundTrip(object):
             (brewtils.models.Event, assert_event_equal, lazy_fixture("bg_event")),
             (brewtils.models.Queue, assert_queue_equal, lazy_fixture("bg_queue")),
             (
-                brewtils.models.Principal,
-                assert_principal_equal,
-                lazy_fixture("bg_principal"),
+                brewtils.models.User,
+                assert_user_equal,
+                lazy_fixture("bg_user"),
             ),
-            (brewtils.models.LegacyRole, assert_role_equal, lazy_fixture("bg_role")),
+            (
+                brewtils.models.UserToken,
+                assert_user_token_equal,
+                lazy_fixture("bg_user_token"),
+            ),
+            (
+                brewtils.models.AliasUserMap,
+                assert_alias_user_map_equal,
+                lazy_fixture("bg_alias_user_map"),
+            ),
+            (brewtils.models.Role, assert_role_equal, lazy_fixture("bg_role")),
+            (
+                brewtils.models.UpstreamRole,
+                assert_upstream_role_equal,
+                lazy_fixture("bg_upstream_role"),
+            ),
             (brewtils.models.Job, assert_job_equal, lazy_fixture("bg_job")),
             (brewtils.models.Job, assert_job_equal, lazy_fixture("bg_cron_job")),
             (brewtils.models.Job, assert_job_equal, lazy_fixture("bg_interval_job")),
@@ -901,6 +1109,21 @@ class TestRoundTrip(object):
                 lazy_fixture("bg_subscriber"),
             ),
             (brewtils.models.Topic, assert_topic_equal, lazy_fixture("bg_topic")),
+            (
+                brewtils.models.StatusInfo,
+                assert_status_info_equal,
+                lazy_fixture("bg_status_info"),
+            ),
+            (
+                brewtils.models.StatusHistory,
+                assert_status_history_equal,
+                lazy_fixture("bg_status_history"),
+            ),
+            (
+                brewtils.models.Replication,
+                assert_replication_equal,
+                lazy_fixture("bg_replication"),
+            ),
         ],
     )
     def test_parsed_start(self, model, assertion, data):
@@ -923,8 +1146,10 @@ class TestRoundTrip(object):
             (brewtils.models.LoggingConfig, lazy_fixture("logging_config_dict")),
             (brewtils.models.Event, lazy_fixture("event_dict")),
             (brewtils.models.Queue, lazy_fixture("queue_dict")),
-            (brewtils.models.Principal, lazy_fixture("principal_dict")),
-            (brewtils.models.LegacyRole, lazy_fixture("legacy_role_dict")),
+            (brewtils.models.User, lazy_fixture("user_dict")),
+            (brewtils.models.UserToken, lazy_fixture("user_token_dict")),
+            (brewtils.models.Role, lazy_fixture("role_dict")),
+            (brewtils.models.UpstreamRole, lazy_fixture("upstream_role_dict")),
             (brewtils.models.Job, lazy_fixture("job_dict")),
             (brewtils.models.Job, lazy_fixture("cron_job_dict")),
             (brewtils.models.Job, lazy_fixture("interval_job_dict")),
@@ -932,6 +1157,9 @@ class TestRoundTrip(object):
             (brewtils.models.Operation, lazy_fixture("operation_dict")),
             (brewtils.models.Runner, lazy_fixture("runner_dict")),
             (brewtils.models.Resolvable, lazy_fixture("resolvable_dict")),
+            (brewtils.models.StatusInfo, lazy_fixture("status_info_dict")),
+            (brewtils.models.StatusHistory, lazy_fixture("status_history_dict")),
+            (brewtils.models.Replication, lazy_fixture("replication_dict")),
         ],
     )
     def test_serialized_start(self, model, data):